+++ index.html
@@ -10,6 +10,7 @@
 </head>
 
 <body>
+
   <main>
 
     <!-- Фильтрация изображений от других пользователей -->
@@ -181,7 +182,6 @@
         <button type="reset" class="big-picture__cancel  cancel" id="picture-cancel">Закрыть</button>
       </div>
     </section>
-    <script src="functions.js"></script>
   </main>
 
   <footer class="page-footer  container">
@@ -227,6 +227,6 @@
       </div>
     </section>
   </template>
-<script src="/js/main.js"></script>
+  <script type="module" src="/js/main.js"></script>
 </body>
 </html>
